requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-jose
bcrypt
requests
httpx
